Reorganized Tomcat integration to separate module.
