+++ impl/src/main/java/org/glassfish/exousia/constraints/WebResourceCollection.java
