+++ impl/src/main/java/org/glassfish/exousia/constraints/transformer/PatternBuilder.java
