+++ impl/src/main/java/org/glassfish/exousia/modules/locked/SharedState.java
