+++ spi/tomcat/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,0 +1,1 @@
+org.glassfish.exousia.spi.tomcat.TomcatIntegrationInitializer
